enhance security with nonces, abspath defined, class_exists, escaping

+++ pages/company_team_delete.php
@@ -1,10 +1,15 @@
 <?php
-// delete current member
 
-global $wpdb;
-$valid = true;
+if (current_user_can('manage_options')) {
+  // delete current member
 
-// prepare get statement protect against SQL inject
-$sql = $wpdb->prepare("DELETE FROM " . $wpdb->prefix . "company_team WHERE id = %d", $id);
+  global $wpdb;
+  $valid = true;
 
-$row = $wpdb->query($sql);
+  // prepare get statement protect against SQL inject
+  $sql = $wpdb->prepare("DELETE FROM " . $wpdb->prefix . "company_team WHERE id = %d", $id);
+
+  $row = $wpdb->query($sql);
+} else {
+  echo 'You are not authorized to perform this action.';
+}
